chore: made requested changes and updated constants
